+++ CHANGELOG.md
@@ -21,7 +21,8 @@
 - Add ability to use a different username instead of hardcoding to 'winesap'.
 - Add support for major upgrades from Mac Linux Gaming Stick 2 to winesapOS 3.
 - Add uninstall script to switch from winesapOS back to upstream Arch Linux.
-- Add support for Mac drivers on Linux LTS 6.1
+- Add support for Mac drivers on Linux LTS 6.1.
+- Add support for installing the GNOME desktop environment.
 
 ## 3.2.1 - The SteamOS 3.4 Update